pyproject.toml
[tool.poetry]
name = "humanitas-poc"
version = "0.1.0"
description = "Proof-of-concept for HUMANITAS academic validation."
authors = ["hawkaiglai <schuks01@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.11"
click = "^8.2.1"
structlog = "^25.4.0"
python-dotenv = "^1.1.0"
numpy = "^2.3.0"
pandas = "^2.3.0"
opencv-python = "^4.11.0"
pillow = "^11.2.1"
scikit-image = "^0.25.2"
scikit-learn = "^1.7.0"
argon2-cffi = "^25.1.0"
face-recognition = "^1.3.0"
dlib = "^20.0.0"
psutil = "^7.0.0"

[tool.poetry.dev-dependencies]
pytest = "^8.4.1"
ruff = "^0.12.0"
black = "^25.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
